LayoutTests/http/tests/cache-storage/cache-clearing-origin.https.html

<!DOCTYPE html>
<html>
<head>
<title>Cache Storage: testing Cache persistency</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <script>
promise_test(test => {
    return self.caches.keys().then(keys => {
        var pending = [];
        for (key of keys)
            pending.push(self.caches.delete(keys[0]));
        return Promise.all(pending);
    });
}, "Cleaning existing caches");

promise_test(async test => {
    var cache = await self.caches.open("test-cache-records-persistency");
    await cache.put("https://example.com/foo", new Response("body", { statusText: "status" }));

    if (!window.testRunner)
        return Promise.reject("test runner needed");

    assert_false(testRunner.hasDOMCache('https://localhost:80'), 'hasDOMCache with fake origin');
    assert_true(testRunner.hasDOMCache(window.location.origin), "hasDOMCache with actual origin");

    testRunner.clearDOMCache('https://localhost:80');

    var keys = await self.caches.keys();
    assert_not_equals(keys.length, 0, "keys should not be empty");
    assert_true(testRunner.domCacheSize(window.location.origin) > 0, "Actual origin cache size is not zero");

    testRunner.clearDOMCache(window.location.origin);
    assert_false(testRunner.hasDOMCache(window.location.origin), "Actual origin cache is cleared");
    assert_equals(testRunner.domCacheSize(window.location.origin), 0, "Actual origin cache size is zero");

    keys = await self.caches.keys();
    assert_equals(keys.length, 0, "keys should be empty");

    cache = await self.caches.open("test-cache-records-persistency");
    keys = await cache.keys();
    assert_equals(keys.length, 0, "records should be empty");
}, 'Clearing disk cache of a given origin');
    </script>
</body>
</html>
